+++ actions/zendesk/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [1.0.2] - 2024-10-04
+## [1.1.0] - 2024-01-03
+
+### Added
+
+- Create a ticket based on the provided attributes
+- Delete ticket with ticket id
 
 ### Changed